You did a lot, here.
Side bar and bottom bar for smaller screens are completed and partially debugged. (The only main thing you saw was that the buttons for the sidebar wouldn't take up the whole width but that's probably because they were buttons rather than anchors.)

Now we need to get the text to float to the right side of the image. Currently the text are three separate paragraphs without a DIV to contain them. Make a DIV to contain them and then the floating should work. Well done!!